#3941 `parse`, `print` tests that use `Farea` to integrations
