+++ .codacy.yml
@@ -6,4 +6,5 @@
 #  package name contains capital letter and such names are conventional.
 ---
 exclude_paths:
-  - "eo-maven-plugin/src/test/java/org/eolang/maven/LintMojoIT.java"
+  - "eo-maven-plugin/src/test/java/org/eolang/maven/ParseMojoIT.java"
+  - "eo-maven-plugin/src/test/java/org/eolang/maven/PrintMojoIT.java"
